# apps/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gun Del Sol API",
    "description": "High-performance async API for Solana token analysis (Modular)",
    "version": "2.0.0"
  },
  "paths": {
...
    "/register": {
      "post": {
        "tags": [
          "Watchlist"
        ],
        "summary": "Register Address",
        "description": "Register a wallet address for monitoring\n\nArgs:\n    payload: Address and optional note\n\nReturns:\n    Registration status and details",
        "operationId": "register_address_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/address/{address}": {
      "get": {
        "tags": [
          "Watchlist"
        ],
        "summary": "Get Address",
        "description": "Get details of a specific monitored address\n\nArgs:\n    address: Solana wallet address\n\nReturns:\n    Address details",
        "operationId": "get_address_address__address__get",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Watchlist"
        ],
        "summary": "Delete Address",
        "description": "Remove an address from monitoring\n\nArgs:\n    address: Solana wallet address\n\nReturns:\n    Deletion status",
        "operationId": "delete_address_address__address__delete",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/address/{address}/note": {
      "put": {
        "tags": [
          "Watchlist"
        ],
        "summary": "Update Address Note",
        "description": "Update the note for a monitored address\n\nArgs:\n    address: Solana wallet address\n    payload: New note text\n\nReturns:\n    Update status",
        "operationId": "update_address_note_address__address__note_put",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressNoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import": {
      "post": {
        "tags": [
          "Watchlist"
        ],
        "summary": "Import Addresses",
        "description": "Import multiple addresses at once\n\nArgs:\n    payload: List of address entries\n\nReturns:\n    Import summary",
        "operationId": "import_addresses_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportAddressesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/multi-token-wallets": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "summary": "Get Multi Early Buyer Wallets",
        "description": "Get wallets that appear in multiple tokens",
        "operationId": "get_multi_early_buyer_wallets_multi_token_wallets_get",
        "parameters": [
          {
            "name": "min_tokens",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2,
              "title": "Min Tokens"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiTokenWalletsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/wallets/refresh-balances": {
      "post": {
        "tags": [
          "Wallets"
        ],
        "summary": "Refresh Wallet Balances",
        "description": "Refresh wallet balances for multiple wallets (ASYNC)",
        "operationId": "refresh_wallet_balances_wallets_refresh_balances_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshBalancesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshBalancesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddressNoteRequest": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "AddressNoteRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportAddressEntry": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          },
          "registered_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registered At"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          },
          "total_notifications": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Notifications"
          },
          "last_notification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Notification"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "address"
        ],
        "title": "ImportAddressEntry"
      },
      "ImportAddressesRequest": {
        "properties": {
          "addresses": {
            "items": {
              "$ref": "#/components/schemas/ImportAddressEntry"
            },
            "type": "array",
            "title": "Addresses"
          }
        },
        "type": "object",
        "required": [
          "addresses"
        ],
        "title": "ImportAddressesRequest"
      },
...
      "MultiTokenWallet": {
        "properties": {
          "wallet_address": {
            "type": "string",
            "title": "Wallet Address"
          },
          "token_count": {
            "type": "integer",
            "title": "Token Count"
          },
          "token_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Token Names"
          },
          "token_addresses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Token Addresses"
          },
          "token_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Token Ids"
          },
          "wallet_balance_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wallet Balance Usd"
          }
        },
        "type": "object",
        "required": [
          "wallet_address",
          "token_count",
          "token_names",
          "token_addresses",
          "token_ids",
          "wallet_balance_usd"
        ],
        "title": "MultiTokenWallet"
      },
      "MultiTokenWalletsResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "wallets": {
            "items": {
              "$ref": "#/components/schemas/MultiTokenWallet"
            },
            "type": "array",
            "title": "Wallets"
          }
        },
        "type": "object",
        "required": [
          "total",
          "wallets"
        ],
        "title": "MultiTokenWalletsResponse"
      },
...
      "RefreshBalancesRequest": {
        "properties": {
          "wallet_addresses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Wallet Addresses"
          }
        },
        "type": "object",
        "required": [
          "wallet_addresses"
        ],
        "title": "RefreshBalancesRequest"
      },
      "RefreshBalancesResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/RefreshBalancesResult"
            },
            "type": "array",
            "title": "Results"
          },
          "total_wallets": {
            "type": "integer",
            "title": "Total Wallets"
          },
          "successful": {
            "type": "integer",
            "title": "Successful"
          },
          "api_credits_used": {
            "type": "integer",
            "title": "Api Credits Used"
          }
        },
        "type": "object",
        "required": [
          "message",
          "results",
          "total_wallets",
          "successful",
          "api_credits_used"
        ],
        "title": "RefreshBalancesResponse"
      },
      "RefreshBalancesResult": {
        "properties": {
          "wallet_address": {
            "type": "string",
            "title": "Wallet Address"
          },
          "balance_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance Usd"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          }
        },
        "type": "object",
        "required": [
          "wallet_address",
          "balance_usd",
          "success"
        ],
        "title": "RefreshBalancesResult"
      },
...
      "RegisterAddressRequest": {
        "properties": {
          "address": {
            "type": "string",
            "maxLength": 44,
            "minLength": 32,
            "title": "Address"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "address"
        ],
        "title": "RegisterAddressRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}